Added DM dir.

+++ dm/pie_host.sql
@@ -1,5 +1,5 @@
 CREATE TABLE pie_host(
-        hst_id BIGINT NOT NULL, 
+        hst_id INTEGER NOT NULL, 
         hst_name VARCHAR(64) NOT NULL,
         hst_fqdn VARCHAR(64) NOT NULL,
         PRIMARY KEY(hst_id),
