+++ dm/pie_mountpoint.sql
@@ -1,6 +1,6 @@
 CREATE TABLE pie_mountpoint(
-        mnt_hst_id BIGINT NOT NULL,
-        mnt_stg_id BIGINT NOT NULL,
+        mnt_hst_id INTEGER NOT NULL,
+        mnt_stg_id INTEGER NOT NULL,
         mnt_path VARCHAR(128) NOT NULL,
         PRIMARY KEY(mnt_hst_id, mnt_stg_id),
         CONSTRAINT fk_mnt_hst FOREIGN KEY (mnt_hst_id)
